Add validation modules for flow and agent configurations
- Implement flow documentation completeness validation to ensure each flow config has corresponding markdown documentation with required autogen markers.
- Add invariant validation functions to check for empty flows and agentless steps in flow configurations.
- Introduce Flow Studio synchronization validation to verify the Flow Studio API is running and serving correct data.
- Create utility flow graph validation to ensure consistency in utility flow specifications.
- Implement frontmatter validation for agent files to check YAML syntax and required fields.
- Add microloop phrase validation to prevent the use of deprecated phrases in flow specs and agent definitions.
- Validate agent prompt sections to ensure required headings are present in agent prompt bodies.
- Introduce RUN_BASE path validation to ensure flow specs use the RUN_BASE placeholder instead of hardcoded paths.
- Validate skills declared in agent frontmatter to ensure corresponding SKILL.md files exist and are correctly formatted.
- Add tests to ensure backward compatibility and validate the functionality of the new modular structure.
