+++ swarm/tools/validation/reporting/__init__.py
@@ -1,0 +1,15 @@
+# swarm/tools/validation/reporting/__init__.py
+"""Reporting and output formatting for validation results."""
+
+from .console_output import print_errors, print_json_output, print_success
+from .json_output import build_detailed_json_output, build_report_json
+from .markdown_output import build_report_markdown
+
+__all__ = [
+    "build_detailed_json_output",
+    "build_report_json",
+    "build_report_markdown",
+    "print_errors",
+    "print_json_output",
+    "print_success",
+]
